added safe migration re-run and fixed issue with elasticsearch host and api host ip in docker and added postman collection and removed unused code

+++ db/migrate/20251110185453_remove_token_and_add_redis_host_to_app.rb
@@ -1,11 +1,11 @@
 class RemoveTokenAndAddRedisHostToApp < ActiveRecord::Migration[5.2]
-  def up 
-    add_column :apps, :redis_host, :integer
-    remove_column :apps, :token
+  def up
+    add_column :apps, :redis_host, :integer, if_not_exists: true
+    remove_column :apps, :token if column_exists?(:apps, :token)
   end
 
-  def down 
-    remove_column :apps, :redis_host
-    add_column :apps, :token, :string
+  def down
+    remove_column :apps, :redis_host if column_exists?(:apps, :redis_host)
+    add_column :apps, :token, :string, if_not_exists: true
   end
 end
